001
When you’re finished creating or making changes to a file on GitHub, scroll to the bottom of the page. Then find the "Commit new file" section.

In the first field, type a commit message. The commit message should briefly tell contributors about the changes you are introducing to the file.

Rules to live by for commit messages:
Don’t end your commit message with a period.
Keep your commit messages to 50 characters or less. Add extra detail in the extended description window if necessary. This is located just below the subject line.
Use active voice. For example, "add" instead of "added" and "merge" instead of "merged".
Think of your commit as expressing intent to introduce a change.